requirements.txt
absl-py==1.0.0
albumentations==1.1.0
astunparse==1.6.3
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.10
colorama==0.4.4
cycler==0.11.0
flatbuffers==2.0
fonttools==4.28.5
gast==0.4.0
grpcio==1.43.0
h5py==3.6.0
idna==3.3
imageio==2.13.5
importlib-metadata==4.10.0
joblib==1.1.0
keras==2.7.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
libclang==12.0.0
Markdown==3.3.6
matplotlib==3.5.1
networkx==2.6.3
numpy==1.22.0
oauthlib==3.1.1
opencv-python==4.5.5.62
opencv-python-headless==4.5.5.62
opt-einsum==3.3.0
packaging==21.3
pandas==1.3.5
Pillow==9.0.0
protobuf==3.19.3
py4j==0.10.9.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.6
python-dateutil==2.8.2
pytz==2021.3
rsa==4.8
scikit-image==0.19.1
scikit-learn==1.0.2
scipy==1.7.3
six==1.16.0
sklearn==0.0
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.7.0
tensorflow-gpu==2.7.0
tensorflow-io-gcs-filesystem==0.23.1
termcolor==1.1.0
threadpoolctl==3.0.0
tifffile==2021.11.2
tqdm==4.62.3
typing_extensions==4.0.1
urllib3==1.26.8
Werkzeug==2.0.2
wrapt==1.13.3
zipp==3.7.0
